Early model testing in chatbot development
Explains the testing of the 'glove-wiki-gigaword-100' embedding model during the early stages of chatbot development. Note that this notebook only contains code for this initial model, and the final chatbot code is available in the Studentia-Chatbot repository. This is not the model we used, though we wanted to share te process we went through.

TL;DR
Added information about the initial challenges with the 'glove-wiki-gigaword-100' embedding model and the decision to switch to OpenAI's 'Curie' models for enhanced efficiency and portability.